remove obsolete dialogs module
